puzzle 2

+++ src/app.ts
@@ -23,7 +23,7 @@ if (!puzzle){
     console.log('Advent Code Breaker! Please pick which number puzzle you\'d like solved')
 } else {
     try {
-        require(`./${parseInt(puzzle)}`)
+        require(`./${parseFloat(puzzle)}`)
     }
     catch {
         console.error('Could not find the puzzle you asked for.')
